Assignment questions
Q1. Which keyword is used to create a function? Create a function to return a list of odd numbers in the
range of 1 to 25.
Q2. Why *args and **kwargs is used in some functions? Create a function each for *args and **kwargs
to demonstrate their use.
Q3. What is an iterator in python? Name the method used to initialise the iterator object and the method
used for iteration. Use these methods to print the first five elements of the given list [2, 4, 6, 8, 10, 12, 14,
16, 18, 20].
Q4. What is a generator function in python? Why yield keyword is used? Give an example of a generator
function.
Q5. Create a generator function for prime numbers less than 1000. Use the next() method to print the
first 20 prime numbers.
Q6. Write a python program to print the first 10 Fibonacci numbers using a while loop.
Q7. Write a List Comprehension to iterate through the given string: ‘pwskills’.
Expected output: ['p', 'w', 's', 'k', 'i', 'l', 'l', 's']
Q8. Write a python program to check whether a given number is Palindrome or not using a while loop.
Q9. Write a code to print odd numbers from 1 to 100 using list comprehension.
Note: Use a list comprehension to create a list from 1 to 100 and use another List comprehension to filter
out odd numbers.