Assignment Questions
Q1. Which function is used to open a file? What are the different modes of opening a file? Explain each mode
of file opening.
Q2. Why close() function is used? Why is it important to close a file?
Q3. Write a python program to create a text file. Write ‘I want to become a Data Scientist’ in that file. Then
close the file. Open this file and read the content of the file.
Q4. Explain the following with python code: read(), readline() and readlines().
Q5. Explain why with statement is used with open(). What is the advantage of using with statement and
open() together?
Q6. Explain the write() and writelines() functions. Give a suitable example.