Assignment Questions
Q1. What is an Exception in Python? Write the difference between Exceptions and syntax errors.
Q2. What happens when an exception is not handled? Explain it with an example.
Q3. Which Python statements are used to catch and handle exception? Explain it with an example.
Q4. Explain with an example
    a. try and else
    b. finally
    c. raise
Q5. What are custom exceptions in Python? Why do we need custom exceptions? Explain with an example.
Q6. Create a custom exception class. Use this class to handle an exception.