Assignment Questions
Q1. What is a database? Differentiate between SQL and NoSQL databases.
Q2. What is DDL? Explain why CREATE, DROP, ALTER, and TRUNCATE are used with an example.
Q3. What is DML? Explain INSERT, UPDATE, and DELETE with an example.
Q4. What is DQL? Explain SELECT with an example.
Q5. Explain Primary Key and Foreign Key.
Q6. Write a Python code to connect MySQL to Python. Explain the cursor() and execute() method.
Q7. Give the order of execution of SQL clauses in an SQL query.