Assignment Questions
Q1. What is Flask Framework? What are the advantages of Flask Framework?
Q2. Create a simple Flask application to display ‘Hello World!!’. Attach the screenshot of the output in
Jupyter Notebook.
Q3. What is App routing in Flask? Why do we use app routes?
Q4. Create a “/welcome” route to display the welcome message “Welcome to ABC Corporation” and a “/”
route to show the following details:
Company Name: ABC Corporation
Location: India
Contact Detail: 999-999-9999
Attach the screenshot of the output in Jupyter Notebook.
Q5. What function is used in Flask for URL Building? Write a Python code to demonstrate the working of the
url_for() function.